Audit Study details

Audit Study details from a 'DataSHIELD' server, an RO-Crate object or a file path pointing to an RO-Crate.

audit_study(x, ...)

# Default S3 method
audit_study(x, ...)

# S3 method for class 'list'
audit_study(
  x,
  ...,
  project = NULL,
  logs_from = -Inf,
  logs_to = Inf,
  path = NULL
)

# S3 method for class 'opal'
audit_study(
  x,
  ...,
  project = NULL,
  logs_from = -Inf,
  logs_to = Inf,
  path = NULL
)

Arguments

x

This can be a connection to a 'DataSHIELD' server (e.g., object with the opal class, see opalr::opal.login()).

...

Other optional arguments, see full documentation for details.

project

String with project name(s) from which to extra Safe People details.

logs_from

Lower limit timestamp to filter out the outputs generated (default: -Inf, everything up to logs_to)

logs_to

Upper limit timestamp to filter out the outputs generated (default: Inf, everything from logs_from onwards).

path

String with path pointing to the root of the RO-Crate. This will be used to store log files. If not provided, logs will be stored within the RO-Crate returned by this function.

Value

Updated RO-Crate object with Study information.