Generate Markdown report's body

Generate Markdown report's body

.markdown_report_body(
  report_contents,
  overview_tbl,
  safe_people_tbl,
  safe_project_tbl,
  safe_data_tbl,
  safe_user_perm_tbl,
  safe_setting_tbl,
  safe_output_tbl,
  break_by = NULL
)

Arguments

report_contents

String with Markdown report (e.g., header).

overview_tbl

Data frame with overview details for the RO-Crate.

safe_people_tbl

Data frame with Safe People details.

safe_project_tbl

Data frame with Safe Project details.

safe_data_tbl

Data frame with Safe Data details.

safe_user_perm_tbl

Data frame with Safe Data user permissions.

safe_setting_tbl

Data frame with Safe Setting details.

safe_output_tbl

Data frame with Safe Output details.

break_by

Optional string with variable to be used for breaking down each table (e.g., server), as opposed to display all the results in a single table.

Value

String with updated Markdown report.