Initialise a Five Safes RO-Crate

Creates a new RO-Crate configured for Five Safes auditing.

init(x, ...)

# S3 method for class 'ArmadilloCredentials'
init(x, ...)

# S3 method for class 'opal'
init(
  x,
  ...,
  rocrate = rocrateR::rocrate_5s(),
  profile = "default",
  project = NULL,
  resources = NULL,
  tables = NULL,
  path = NULL,
  user = NULL
)

# S3 method for class 'rocrate'
init(
  x,
  ...,
  connection = attr(x, "connection"),
  path = attr(x, "path"),
  profile = attr(x, "profile"),
  project = attr(x, "project"),
  resources = attr(x, "resources"),
  tables = attr(x, "tables"),
  user = attr(x, "user")
)

Arguments

x

This can be a connection to a 'DataSHIELD' server (e.g., object with the opal or ArmadilloCredentials classes), an RO-Crate (rocrate class) or a string with the path to an RO-Crate.

...

Other optional arguments. See the full documentation, ?dsROCrate::init.

rocrate

RO-Crate object. Optional, if x is either an RO-Crate object or a path to a valid RO-Crate. If so, then connection is required (default: rocrateR::rocrate_5s()).

profile

String with profile name (used for OBiBa's Opal backend).

project

String with the name of the Safe Project.

resources

Vector of strings with the names of the resources, part of project. Optional, if not provided, all the resources associated to project will be included in the RO-Crate.

tables

Vector of strings with the names of the tables/datasets, part of project. Optional, if not provided, all the tables/datasets associated to project will be included in the RO-Crate.

path

String with path pointing to the root of the RO-Crate. This will be used to store log files. If not provided, logs will be stored within the RO-Crate returned by this function.

user

List (or entity object) with details for the Safe People, it must include @id and name entries. Alternatively, this can be a string with the name of the current user.

connection

Connection object for the 'DataSHIELD' server where the values will be extracted from (e.g., OBiBa's Opal). Optional, if x is set to a connection object. If so, then rocrate is required.

Value

Five Safes RO-Crate object.

References

Wilkinson, M., Dumontier, M., Aalbersberg, I. et al. (2016) The FAIR Guiding Principles for scientific data management and stewardship. Sci Data 3, 160018. https://doi.org/10.1038/sdata.2016.18