Safe outputs details

Safe outputs details for the RO-Crate.

safe_output(x, ...)

# S3 method for class 'opal'
safe_output(
  x,
  ...,
  rocrate = NULL,
  path = NULL,
  user = NULL,
  logs_to = Sys.time(),
  logs_from = logs_to - 24 * 60^2
)

# S3 method for class 'rocrate'
safe_output(
  x,
  ...,
  path = NULL,
  user = NULL,
  logs_to = Sys.time(),
  logs_from = logs_to - 24 * 60^2,
  connection = NULL
)

Source

Research Data Scotland, 2025. "What is the Five Safes framework?". https://www.researchdata.scot/engage-and-learn/data-explainers/what-is-the-five-safes-framework/

Arguments

x

This can be a connection to a 'DataSHIELD' server (e.g., object with the opal class, see opalr::opal.login()), an RO-Crate (rocrate class) or a string with the path to an RO-Crate.

...

Other optional arguments. See the full documentation, ?dsROCrate::safe_output.

rocrate

RO-Crate object. Optional, if x is either an RO-Crate object or a path to a valid RO-Crate. If so, then connection is required.

path

String to path pointing to the root of the RO-Crate. This will be used to store log files. If not provided, logs will be stored within the updated RO-Crate returned by this function.

user

String with the name of the user for which outputs will be extracted. Optional, if no user is provided, then this will be extracted from the author in the root (./) entity of the RO-Crate.

logs_to

Upper limit timestamp to filter outputs to be extracted (default: Sys.time(), current system time).

logs_from

Lower limit timestamp to filter outputs to be extracted (default: Sys.time() - 24 * 60 ^ 2, last 24 hours)

connection

Connection object for the 'DataSHIELD' server where the values will be extracted from (e.g., OBiBa's Opal). Optional, if x is set to a connection object. If so, then rocrate is required.

Value

Updated RO-Crate object with Safe Outputs information.

Details

All research outputs are checked to ensure individuals cannot be identified even in the public domain.

This means the other four Safes no longer apply.

Before outputs are released from the TRE, they are checked to make sure it is reasonably unlikely that any individuals can be identified on publication. Compliance with the Digital Economy Act also requires that the TRE applies methods and standards for output checking that are accredited by the UK Statistics Authority.