Safe projects details

Safe projects details for the RO-Crate.

# S4 method for class 'armadillo'
safe_project(
  x,
  ...,
  project = NULL,
  rocrate = rocrateR::rocrate_5s(),
  dataset_id_suffix = "#dataset:",
  project_id_suffix = "#project:",
  path = NULL,
  tables = NULL,
  user = NULL
)

# S3 method for class 'character'
safe_project(
  x,
  ...,
  project = attr(x, "project"),
  dataset_id_suffix = "#dataset:",
  project_id_suffix = "#project:",
  connection = attr(x, "connection"),
  path = attr(x, "path"),
  tables = attr(x, "tables"),
  user = attr(x, "user")
)

# S3 method for class 'opal'
safe_project(
  x,
  ...,
  project = NULL,
  rocrate = rocrateR::rocrate_5s(),
  dataset_id_suffix = "#dataset:",
  project_id_suffix = "#project:",
  path = NULL,
  tables = NULL,
  user = NULL
)

# S3 method for class 'rocrate'
safe_project(
  x,
  ...,
  project = attr(x, "project"),
  dataset_id_suffix = "#dataset:",
  project_id_suffix = "#project:",
  connection = attr(x, "connection"),
  path = attr(x, "path"),
  tables = attr(x, "tables"),
  user = attr(x, "user")
)

Source

Research Data Scotland, 2025. "What is the Five Safes framework?". https://www.researchdata.scot/engage-and-learn/data-explainers/what-is-the-five-safes-framework/

Arguments

x

This can be a connection to a 'DataSHIELD' server (e.g., object with the opal class, see opalr::opal.login()), an RO-Crate (rocrate class) or a string with the path to an RO-Crate.

...

Other optional arguments. See the full documentation, ?dsROCrate::safe_project.

project

String with the name of the Safe Project.

dataset_id_suffix

String with ID suffix for the tables/datasets entities in the RO-Crate (default: "#dataset:").

project_id_suffix

String with ID suffix for the project entities in the RO-Crate (default: "#project:").

connection

Connection object for the 'DataSHIELD' server where the values will be extracted from (e.g., OBiBa's Opal). Optional, if x is set to a connection object. If so, then rocrate is required.

path

String with path pointing to the root of the RO-Crate. This will be used to store log files. If not provided, logs will be stored within the RO-Crate returned by this function.

tables

Vector of strings with the names of the tables/datasets, part of project. Optional, if not provided, all the tables/datasets associated to project will be included in the RO-Crate.

user

List (or entity object) with details for the Safe People, it must include @id and name entries. Alternatively, this can be a string with the name of the current user.

rocrate

RO-Crate object. Optional, if x is either an RO-Crate object or a path to a valid RO-Crate. If so, then connection is required (default: rocrateR::rocrate_5s()).

Value

Updated RO-Crate object with Safe Projects information.

Details

Data must be used ethically, for research that delivers clear public benefit.

As part of their application, researchers are asked to provide an overview of their project, including how the data will be used and what outputs will be achieved. This allows data providers to make an informed decision about whether they are comfortable preparing data for the researcher to use for ethical purposes serving a public good.